+++ src/file/file_macho.cc
@@ -218,14 +218,14 @@ static void file_load_macho(struct machine *m, struct memory *mem,
 
 			CHECK_ALLOCATION(symbols = (char *) malloc(12 * nsyms));
 			fseek(f, symoff, SEEK_SET);
-			if (fread(symbols, 1, 12 * nsyms, f) != 12*nsyms) {
+			if (fread(symbols, 1, 12 * nsyms, f) != (size_t) 12*nsyms) {
 				fprintf(stderr, "could not read symbols from %s\n", filename);
 				exit(1);
 			}
 
 			CHECK_ALLOCATION(strings = (char *) malloc(strsize));
 			fseek(f, stroff, SEEK_SET);
-			if (fread(strings, 1, strsize, f) != strsize) {
+			if (fread(strings, 1, strsize, f) != (size_t) strsize) {
 				fprintf(stderr, "could not read symbol strings from %s\n", filename);
 				exit(1);
 			}
